api/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
pydantic-settings
python-dotenv
httpx
playwright
beautifulsoup4
lxml
resend
